feat(workbench,desktop): share provider × account settings editor
