+++ packages/workbench/src/index.ts
@@ -29,6 +29,7 @@ export * from './runtime/tayori';
 export * from './scripts/hooks';
 export * from './scripts/open-url-store';
 export * from './scripts/services-menu';
+export * from './settings/agent-accounts';
 export * from './surface/selection-store';
 export * from './surface/shell';
 export * from './surface/use-workbench-sessions';
